chores: move CreateGatheringBtn into Search directory
